Renamed mockData folder.
