+++ src/pages/Calendar/Calendar.test.tsx
@@ -1,7 +1,7 @@
 import { screen } from "@testing-library/react";
 import Calendar from "./Calendar";
 import { customRender } from "../../utils/testUtils";
-import { CompletedTask } from "../../tempname/mockCompletedTasks";
+import { CompletedTask } from "../../mockData/mockCompletedTasks";
 import dayjs from "dayjs";
 import updateLocale from "dayjs/plugin/updateLocale";
